Create a dedicated dex-testing module
Currently the jmh benchmarks have access to the `randomizedrunner`
dependencies, which cause failures with `Randomness.get()` calls: They
fail because the `RandomizedContext` is available in the classpath but
not setup correctly, since jmh benchmarks do not run in a
`RandomizedRunner`.

This is a first step of a module restructure, to eventually get rid of
the `randomizedrunner` dependencies from the jmh benchmarks.